# peformance improvement for packingReader function