packingReader is the most time-consuming part in Query API.
this tiny PR improves the query response time for a 1Sec-timeframe bucket by 10%.

This branch:
```
$ msbench --write-num=1 --query-num=100 --timeframes=1Sec --range=1year --limit=""
timeframe, elapsed_time_per_operation(write/query)
[is_variable_length=False, data_size=100]
1Sec   24.13967ms/write
1Sec   137.91687ms/query
[is_variable_length=True, data_size=100]
1Sec   27.35185ms/write
1Sec   150.87722ms/query
```

master branch:
```
$ msbench --write-num=1 --query-num=100 --timeframes=1Sec --range=1year --limit=""
timeframe, elapsed_time_per_operation(write/query)
[is_variable_length=False, data_size=100]
1Sec   25.22351ms/write
1Sec   157.51265ms/query
[is_variable_length=True, data_size=100]
1Sec   27.45766ms/write
1Sec   170.08065ms/query
```
cf. benchmarking tool I made for marketstore: https://github.com/dakimura/msbench